ADDED: Signals support
Signals are groups of nets that are connected by a single component.
When defined as a signal, ERC will test compatibility of pins on the
signal.  PCB Editor allows tuning signal length.  Also provides
dedicated custom DRC properties and constraints for signals vs nets.

Signals can be named and referenced in design rules